requirements.txt
psycopg2-binary==2.8.6
pylama==7.7.1